+++ client/src/main/java/com/vesoft/nebula/graph/client/async/AsyncGraphClient.java
@@ -6,13 +6,7 @@
 
 package com.vesoft.nebula.graph.client.async;
 
-import com.facebook.thrift.TException;
-import com.google.common.base.Optional;
-import com.google.common.util.concurrent.ListenableFuture;
 import com.vesoft.nebula.Client;
-import com.vesoft.nebula.graph.client.ConnectionException;
-import com.vesoft.nebula.graph.client.NGQLException;
-import com.vesoft.nebula.graph.client.ResultSet;
 import com.vesoft.nebula.graph.client.async.entry.ExecuteCallback;
