--- client/src/main/java/com/vesoft/nebula/graph/client/async/NebulaResultCodeFuture.java
@@ -1,13 +1,0 @@
-/* Copyright (c) 2019 vesoft inc. All rights reserved.
- *
- * This source code is licensed under Apache 2.0 License,
- * attached with Common Clause Condition 1.0, found in the LICENSES directory.
- */
-
-package com.vesoft.nebula.graph.client.async;
-
-import com.google.common.util.concurrent.AbstractFuture;
-import com.vesoft.nebula.graph.ErrorCode;
-
-public class NebulaResultCodeFuture extends AbstractFuture<ErrorCode> {
-}
