+++ examples/src/main/java/com/vesoft/nebula/examples/AsyncMetaClientExample.java
@@ -4,29 +4,19 @@
  * attached with Common Clause Condition 1.0, found in the LICENSES directory.
  */
 
-package com.vesoft.nebula.meta.client.async;
+package com.vesoft.nebula.examples;
 
 import com.facebook.thrift.TBase;
 import com.facebook.thrift.TException;
 import com.google.common.base.Optional;
-import com.google.common.net.HostAndPort;
-import com.google.common.util.concurrent.FutureCallback;
-import com.google.common.util.concurrent.Futures;
-import com.google.common.util.concurrent.ListenableFuture;
-import com.google.common.util.concurrent.ListeningExecutorService;
-import com.google.common.util.concurrent.MoreExecutors;
 import com.vesoft.nebula.meta.ErrorCode;
 import com.vesoft.nebula.meta.IdName;
 import com.vesoft.nebula.meta.ListSpacesResp;
+import com.vesoft.nebula.meta.client.async.AsyncMetaClientImpl;
 import com.vesoft.nebula.meta.client.async.entry.ListSpaceCallback;
-import com.vesoft.nebula.meta.client.entry.GetPartsAllocResult;
 import com.vesoft.nebula.meta.client.entry.ListSpaceResult;
 
-import java.util.List;
 import java.util.Map;
-import java.util.concurrent.Executors;
-
-import org.checkerframework.checker.nullness.qual.Nullable;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -37,8 +27,8 @@ public class AsyncMetaClientExample {
 
     public static void main(String[] args) {
         if (args.length != 2) {
-            System.out.println("Usage: " + "com.vesoft.nebula.examples.AsyncMetaClientExample " +
-                "<host> " + "<port>");
+            System.out.println("Usage: " + "com.vesoft.nebula.examples.AsyncMetaClientExample "
+                + "<host> " + "<port>");
             return;
         }
         AsyncMetaClientImpl asyncMetaClient = new AsyncMetaClientImpl(args[0],
@@ -77,38 +67,5 @@ public class AsyncMetaClientExample {
             LOGGER.error(e.getMessage());
             e.printStackTrace();
         }
-
-
-        /*
-        ListenableFuture<Optional<GetPartsAllocResult>> getPartsAllocResult =
-                asyncMetaClient.getPartsAlloc(1);
-        Futures.addCallback(
-                getPartsAllocResult,
-                new FutureCallback<Optional<GetPartsAllocResult>>() {
-                    @Override
-                    public void onSuccess(
-                            @Nullable Optional<GetPartsAllocResult> getPartsAllocResultOptional) {
-                        if (getPartsAllocResultOptional.isPresent()) {
-                            GetPartsAllocResult result = getPartsAllocResultOptional.get();
-                            Map<Integer, List<HostAndPort>> map = result.getResult();
-                            LOGGER.info("---------------Spaces:--------------");
-                            for (Map.Entry<Integer, List<HostAndPort>> entry : map.entrySet()) {
-                                LOGGER.info(String.format("Part ID: %d, ", entry.getKey())
-                                        + String.format("Addresses: %s", entry.getValue()));
-                            }
-                            LOGGER.info("------------------------------------");
-                        } else {
-                            LOGGER.info(String.format("No Part Founded"));
-                        }
-                    }
-
-                    @Override
-                    public void onFailure(Throwable throwable) {
-                        LOGGER.error("Get Parts Failed");
-                    }
-                },
-                EXECUTOR_SERVICE);
-
-         */
     }
 }
